TGIS-master/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-ratelimit==4.1.0
django-widget-tweaks==1.5.0
simplekml==1.3.6
fastkml==1.1.0
geopy==2.4.1
lxml==6.0.0
Pillow==11.3.0
weasyprint==65.1
xlsxwriter==3.2.5
qrcode==8.2
python-decouple==3.8
arrow==1.3.0
pygeoif==1.5.1
geographiclib==2.0
pandas
simplekml
geopandas
python-magic
